requirements.txt
playwright
selectolax
python-dotenv
icecream
typer
xlwings